Correcoes estruturais FK e paginas

+++ src/main/java/com/gestaotreinamento/repository/SalaRepository.java
@@ -11,13 +11,13 @@ import com.gestaotreinamento.model.Sala;
 
 @Repository
 @Transactional
-public interface SalaRepository extends CrudRepository<Sala, Long> {
+public interface SalaRepository extends CrudRepository<Sala, Integer> {
 
 	@Query("select count(id) from Sala s")
 	Integer findTotalSalas();
 	
 	@Query("select s.id from Sala s")
-	List<Long> findAllId();
+	List<Integer> findAllId();
 	
 	@Query("select p from Sala p where nomeSala like %?1%")
 	List<Sala> findSalaPorNome(String nomeSala);
